Move assume-valid hashes to chainparams
