+++ crates/floresta-chain/src/pruned_utreexo/chainparams.rs
@@ -16,6 +16,7 @@ use bitcoin::Target;
 use rustreexo::accumulator::node_hash::NodeHash;
 
 use crate::prelude::*;
+use crate::AssumeValidArg;
 use crate::Network;
 #[derive(Clone, Debug)]
 pub struct ChainParams {
@@ -97,7 +98,7 @@ pub struct AssumeUtreexoValue {
 }
 
 impl ChainParams {
-    pub fn get_assumeutreexo_value(network: Network) -> AssumeUtreexoValue {
+    pub fn get_assume_utreexo(network: Network) -> AssumeUtreexoValue {
         match network {
             Network::Bitcoin => AssumeUtreexoValue {
                 block_hash: BlockHash::from_str(
@@ -149,9 +150,35 @@ impl ChainParams {
             },
         }
     }
-}
 
-impl ChainParams {
+    pub fn get_assume_valid(network: Network, arg: AssumeValidArg) -> Option<BlockHash> {
+        fn get_hash(hash: &str) -> BlockHash {
+            BlockHash::from_str(hash).expect("hardcoded hash should not fail")
+        }
+        match arg {
+            AssumeValidArg::Disabled => None,
+            AssumeValidArg::UserInput(hash) => Some(hash),
+            AssumeValidArg::Hardcoded => match network {
+                Network::Bitcoin => {
+                    get_hash("00000000000000000000569f4d863c27e667cbee8acc8da195e7e5551658e6e9")
+                        .into()
+                }
+                Network::Testnet => {
+                    get_hash("000000000000001142ad197bff16a1393290fca09e4ca904dd89e7ae98a90fcd")
+                        .into()
+                }
+                Network::Signet => {
+                    get_hash("0000003ed17b9c93954daab00d73ccbd0092074c4ebfc751c7458d58b827dfea")
+                        .into()
+                }
+                Network::Regtest => {
+                    get_hash("0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206")
+                        .into()
+                }
+            },
+        }
+    }
+
     fn max_target(net: Network) -> Target {
         match net {
             Network::Bitcoin => Target::MAX_ATTAINABLE_MAINNET,
